WebSockets/tests/submissions/Opera/interfaces/WebSocket/readyState/005.html

<!doctype html>
<title>WebSockets: defineProperty setter for readyState</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(assert_throws(new typeError(), function(){
  Object.defineProperty(WebSocket.prototype, 'readyState', {
    set: function() { return 'foo'; }
  });
})
);
</script>
